Move types and interfaces from .slice to separate file
