+++ packages/nectar/src/sagas/communities/launchRegistrar/launchRegistrar.saga.test.ts
@@ -1,18 +1,17 @@
+import { combineReducers } from '@reduxjs/toolkit'
 import { expectSaga } from 'redux-saga-test-plan'
 import { Socket } from 'socket.io-client'
-import { SocketActionTypes } from '../../socket/const/actionTypes'
 import { identityAdapter } from '../../identity/identity.adapter'
-import { launchRegistrarSaga } from './launchRegistrar.saga'
-import { combineReducers } from '@reduxjs/toolkit'
+import { identityReducer, IdentityState } from '../../identity/identity.slice'
+import { Identity } from '../../identity/identity.types'
+import { SocketActionTypes } from '../../socket/const/actionTypes'
 import { StoreKeys } from '../../store.keys'
+import { communitiesAdapter } from '../communities.adapter'
 import {
   communitiesActions,
-  communitiesReducer,
-  Community,
-  CommunitiesState
+  communitiesReducer, CommunitiesState, Community
 } from '../communities.slice'
-import { communitiesAdapter } from '../communities.adapter'
-import { Identity, identityReducer, IdentityState } from '../../identity/identity.slice'
+import { launchRegistrarSaga } from './launchRegistrar.saga'
 
 describe('launchRegistrar', () => {
   test("launch certain registrar instead of current community's registrar", async () => {
