+++ packages/nectar/src/sagas/identity/identity.adapter.ts
@@ -1,5 +1,5 @@
 import { createEntityAdapter } from '@reduxjs/toolkit'
-import { Identity } from './identity.slice'
+import { Identity } from './identity.types'
 
 export const identityAdapter = createEntityAdapter<Identity>({
   selectId: (identity) => identity.id
