+++ packages/nectar/src/sagas/identity/identity.slice.ts
@@ -1,106 +1,16 @@
 import { createSlice, EntityState, PayloadAction } from '@reduxjs/toolkit'
-import { identityAdapter } from './identity.adapter'
 import { StoreKeys } from '../store.keys'
+import { identityAdapter } from './identity.adapter'
+import {
+  CreateUserCsrPayload,
+  Identity,
+  StoreUserCertificatePayload,
+  StoreUserCsrPayload,
+  UpdateUsernamePayload
+} from './identity.types'
 
 export class IdentityState {
   public identities: EntityState<Identity> = identityAdapter.getInitialState()
-}
-
-export interface Identity {
-  id: string
-  zbayNickname: string
-  hiddenService: HiddenService
-  dmKeys: DmKeys
-  peerId: PeerId
-  userCsr: UserCsr | null
-  userCertificate: string | null
-}
-
-export interface CertData {
-  publicKey: any
-  privateKey: any
-  pkcs10: any
-}
-
-export interface UserCsr {
-  userCsr: string
-  userKey: string
-  pkcs10: CertData
-}
-
-export interface CreateDmKeyPairPayload {
-  dmPublicKey: string
-  dmPrivateKey: string
-}
-
-export interface HiddenService {
-  onionAddress: string
-  privateKey: string
-}
-
-export interface PeerId {
-  id: string
-  pubKey?: string
-  privKey?: string
-}
-
-export interface DmKeys {
-  publicKey: string
-  privateKey: string
-}
-
-export interface CreateUserCsrPayload {
-  zbayNickname: string
-  commonName: string
-  peerId: string
-  dmPublicKey: string
-  signAlg: string
-  hashAlg: string
-}
-
-export interface UpdateUsernamePayload {
-  communityId: string
-  nickname: string
-}
-
-export interface RegisterUserCertificatePayload {
-  id: string
-  userCsr: string
-  serviceAddress: string
-}
-
-export interface PermsData {
-  certificate: string
-  privKey: string
-}
-
-export interface RegisterOwnerCertificatePayload {
-  id: string
-  userCsr: string
-  permsData: PermsData
-}
-
-export interface SaveCertificatePayload {
-  certificate: string
-  rootPermsData: PermsData
-}
-
-export interface SaveOwnerCertificatePayload {
-  id: string
-  peerId: string
-  certificate: string
-  permsData: PermsData
-}
-
-export interface StoreUserCertificatePayload {
-  userCertificate: string
-  communityId: string
-}
-
-export interface StoreUserCsrPayload {
-  userCsr: UserCsr
-  communityId: string
-  registrarAddress: string
 }
 
 export const identitySlice = createSlice({
