+++ packages/nectar/src/sagas/identity/saveOwnerCertToDb/saveOwnerCertToDb.saga.ts
@@ -3,7 +3,7 @@ import { apply, select } from 'typed-redux-saga'
 import { SocketActionTypes } from '../../socket/const/actionTypes'
 import { communitiesSelectors } from '../../communities/communities.selectors'
 import { identitySelectors } from '../identity.selectors'
-import { SaveOwnerCertificatePayload } from '../identity.slice'
+import { SaveOwnerCertificatePayload } from '../identity.types'
 
 export function* saveOwnerCertToDbSaga(socket: Socket): Generator {
   const currentCommunity = yield* select(communitiesSelectors.currentCommunity)
