+++ packages/nectar/src/sagas/users/users.types.ts
@@ -1,0 +1,10 @@
+export interface User {
+  username: string
+  onionAddress: string
+  peerId: string
+  dmPublicKey: string
+}
+
+export interface SendCertificatesResponse {
+  certificates: string[]
+}
